chore: delete unused AppNav
