+++ apps/loan/src/layout/index.tsx
@@ -9,10 +9,9 @@ import { isFailure, isLoading } from '@/ui/utils'
 import { useConnectWallet } from '@/common/features/connect-wallet'
 import useLayoutHeight from '@/hooks/useLayoutHeight'
 import useStore from '@/store/useStore'
-
 import Header from '@/layout/Header'
 import GlobalBanner from '@/ui/Banner'
-import { Locale } from '@/ui/AppNav/types'
+import { Locale } from '@/common/widgets/Header/types'
 import { t } from '@lingui/macro'
 import Footer from '@/ui/Footer'
 import networks from '@/networks'
