+++ packages/curve-common/src/widgets/Header/SidebarSection.tsx
@@ -1,7 +1,7 @@
 import type { FunctionComponent, ReactNode } from 'react'
 import { List, ListSubheader } from '@mui/material'
 import { SidebarItem } from './SidebarItem'
-import type { AppPage } from 'ui/src/AppNav/types'
+import type { AppPage } from 'curve-common/src/widgets/Header/types'
 import Divider from '@mui/material/Divider'
 import { Typography } from 'curve-ui-kit/src/shared/ui/Typography'
