--- packages/ui/src/AppNav/index.ts
@@ -1,4 +1,0 @@
-export * from './styles'
-export * from './constants'
-export { default as AppNavMobile } from './AppNavMobile'
-export { default as AppNavSecondary } from './AppNavSecondary'
