--- packages/ui/src/HeaderStats/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HeaderStats'
