+++ packages/ui/src/index.tsx
@@ -4,7 +4,6 @@ export * from './images'
 
 export * from './Accordion'
 export * from './AlertBox'
-export * from './AppNav'
 export * from './AppPage'
 export * from './AppForm'
 export * from './Box'
@@ -20,7 +19,6 @@ export * from './DialogComboBox'
 export * from './DividerHorizontal'
 export * from './Footer'
 export * from './FormErrorsDisplay'
-export * from './HeaderStats'
 export * from './Icon'
 export * from './IconButton'
 export * from './InputComp'
